Refactored the project to modulize view components and filters.
