+++ src/filter/fadeOut.js
@@ -1,0 +1,15 @@
+export default class FadeOut {
+    constructor() {
+
+    };
+
+    static giveEffect(selectedTrackBuffer, startPositionInBuffer, endPositionInBuffer, params) {
+        let regionLengthInBuffer = endPositionInBuffer - startPositionInBuffer;
+        for (var channelNumber = 0; channelNumber < selectedTrackBuffer.numberOfChannels; channelNumber++){
+            var channelData = selectedTrackBuffer.getChannelData(channelNumber);
+            for (var cursor = startPositionInBuffer; cursor < endPositionInBuffer; cursor++){
+                channelData[cursor] *= (endPositionInBuffer - cursor) / regionLengthInBuffer;
+            }
+        }
+    }
+}
