+++ src/index.js
@@ -1,9 +1,22 @@
+import $ from 'jquery';
+import 'dist/css/bootstrap.css';
+import 'dist/css/bootstrap-toggle.css';
+import 'dist/js/bootstrap.min.js';
+import 'dist/js/bootstrap-toggle.min.js';
 import 'css/index.css';
+
+
+
+
+
 import WaveList from 'waveList.js';
+import WaveListModifier from 'waveListModifier';
+
 
 
 $(document).ready(function() {
     let waveList = WaveList.create({});
+    let waveListModifier = WaveListModifier.create(waveList);
 
     /* Test codes for dev, should be erased in production*/
     waveList.addWaveForm(0);
