+++ views/index.html
@@ -42,7 +42,7 @@
           </div>
         </nav>
 
-        <div class="container-fluid">
+        <div class="container">
             
             <div id="waveContent" class="waveform-body" >
                 <div class="row vertical-align-bottom">
